+++ eotsmanager/service/rpcserver.go
@@ -64,6 +64,25 @@ func (r *rpcServer) CreateRandomnessPairList(_ context.Context, req *proto.Creat
 	}, nil
 }
 
+// CreateRandomnessPairListWithInterval returns a list of Schnorr randomness pairs with a specified interval between heights.
+func (r *rpcServer) CreateRandomnessPairListWithInterval(_ context.Context, req *proto.CreateRandomnessPairListWithIntervalRequest) (
+	*proto.CreateRandomnessPairListWithIntervalResponse, error) {
+	pubRandList, err := r.em.CreateRandomnessPairListWithInterval(req.Uid, req.ChainId, req.StartHeight, req.Num, req.Interval)
+
+	if err != nil {
+		return nil, fmt.Errorf("failed to create randomness pair list with interval: %w", err)
+	}
+
+	pubRandBytesList := make([][]byte, 0, len(pubRandList))
+	for _, p := range pubRandList {
+		pubRandBytesList = append(pubRandBytesList, p.Bytes()[:])
+	}
+
+	return &proto.CreateRandomnessPairListWithIntervalResponse{
+		PubRandList: pubRandBytesList,
+	}, nil
+}
+
 // SignEOTS signs an EOTS with the EOTS private key and the relevant randomness
 func (r *rpcServer) SignEOTS(_ context.Context, req *proto.SignEOTSRequest) (
 	*proto.SignEOTSResponse, error) {